django-backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=21.2.0
whitenoise>=6.6.0
